+++ slave_test.go
@@ -91,7 +91,7 @@ func TestTntBoxGuestWrite(t *testing.T) {
 	tnt, err := Connect(box.Listen, &Options{})
 	require.NoError(err)
 	defer tnt.Close()
-	tuple := []interface{}{int8(2), "Client inserted #2"}
+	tuple := []interface{}{int64(2), "Client inserted #2"}
 	res, err := tnt.Execute(&Insert{
 		Space: "tester",
 		Tuple: tuple,
@@ -589,7 +589,7 @@ func TestSlaveAttachAsync(t *testing.T) {
 	tnt, err := Connect(box.Listen, &Options{})
 	require.NoError(err)
 	defer tnt.Close()
-	expected := []interface{}{uint64(2), "Client inserted #2"}
+	expected := []interface{}{int64(2), "Client inserted #2"}
 	res, err := tnt.Execute(&Insert{
 		Space: "tester",
 		Tuple: expected,
